ai/requirements.txt
# AI Service Dependencies
# Core dependencies for processing files with OpenRouter API

# HTTP requests for API calls
requests>=2.31.0

# PDF text extraction
PyPDF2>=3.0.0

# DOCX text extraction
python-docx>=1.1.0

# Environment variable management (optional but recommended)
python-dotenv>=1.0.0

